feat(logout): update tests
