+++ .eslintrc.json
@@ -13,6 +13,7 @@
     "consistent-return": 0,
     "no-param-reassign": 0,
     "comma-dangle": 0,
+    "linebreak-style": 0,
     "import/prefer-default-export": 0,
     "curly": ["error", "multi-line"],
     "import/no-unresolved": [2, { "commonjs": true }],
